requirements.txt
beautifulsoup4
certifi
charset-normalizer
google
googlesearch-python
idna
langchain
numpy
openai
requests
soupsieve
streamlit
tqdm
urllib3
wikipedia
python-dotenv
webscout
asyncio
selenium
platform
zipfile
io
pyperclip